Improve speed calculate code.

+++ src/download/play.go
@@ -24,7 +24,7 @@ func Play(t *task.Task, w io.Writer, from, to int64) {
 
 	play_quit = make(chan bool)
 
-	control := make(chan int)
+	control := make(chan int64)
 	progress := doDownload(t.URL, w, from, to, 0, control, play_quit)
 	handleProgress(progress, t, play_quit)
 }
